Keyboard plug'n'play, cleanup,updated Readme
