--- src/uspi-dummy.cpp
@@ -1,3 +1,0 @@
-#include "rpi-gpio.h"
-
-rpi_gpio_t* RPI_GpioBase = (rpi_gpio_t*) RPI_GPIO_BASE;
